requirements.txt
Jinja2==2.7.2
MarkupSafe==0.19
Pygments==1.6
Sphinx==1.2.2
beautifulsoup4==4.3.2
dnspython==1.11.1
docutils==0.11
gevent==1.0
greenlet==0.4.2
lxml==3.3.4
nose==1.3.1
python-dateutil==2.2
simplejson==3.4.0
six==1.6.1
sleekxmpp==1.2.4
wsgiref==0.1.2
